requirements-dev.txt
codespell
editorconfig-checker
flake8
isort
jupyter
mypy
pylint
pytest
pytest-cov
sphinx
sphinx-autobuild
sphinx-click
sphinxcontrib-fulltoc
sphinxcontrib-napoleon
toml
types-click
types-python-dateutil
types-pytz
yapf
